Make virtio bindings features mutually exclusive
If we are compiling both bindings versions at the same time
(with --all-features), we are exporting two files with the same
name (e.g. virtio_blk) at the same time. Because of that we are
getting warnings at compile time. We should compile/export only
the virtio-v5_0_0 features if --all-features flag is present.

Signed-off-by: Laura Loghin <lauralg@amazon.com>